Fix issue #4114: Remove incomplete project if import process fail
