+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ## 2.1.0 ??? (unreleased)
 
 ### Features
+- Add sprint name and slug on search results for user stories ((thanks to [@everblut](https://github.com/everblut)))
+- [API] projects resource: Random order if `discover_mode=true` and `is_featured=true`.
 - Webhooks: Improve webhook data:
     - add permalinks
     - owner, assigned_to, status, type, priority, severity, user_story, milestone, project are objects
@@ -15,8 +17,9 @@
 - CSV Reports:
     - Change field name: 'milestone' to 'sprint'
     - Add new fields: 'sprint_estimated_start' and 'sprint_estimated_end'
-- Add sprint name and slug on search results for user stories ((thanks to [@everblut](https://github.com/everblut)))
-- [API] projects resource: Random order if `discover_mode=true` and `is_featured=true`.
+- Importer:
+    - Remove project after load a dump file fails
+    - Add more info the the logger if load a dump file fails
 
 ### Misc
 - Lots of small and not so small bugfixes.
