+++ taiga/export_import/api.py
@@ -36,14 +36,14 @@ from taiga.projects.models import Project, Membership
 from taiga.projects.issues.models import Issue
 from taiga.projects.tasks.models import Task
 from taiga.projects.serializers import ProjectSerializer
-from taiga.users import services as users_service
+from taiga.users import services as users_services
 
+from . import exceptions as err
 from . import mixins
-from . import serializers
-from . import service
 from . import permissions
+from . import serializers
+from . import services
 from . import tasks
-from . import dump_service
 from . import throttling
 from .renderers import ExportRenderer
 
@@ -72,7 +72,7 @@ class ProjectExporterViewSet(mixins.ImportThrottlingPolicyMixin, GenericViewSet)
         path = "exports/{}/{}-{}.json".format(project.pk, project.slug, uuid.uuid4().hex)
         storage_path = default_storage.path(path)
         with default_storage.open(storage_path, mode="w") as outfile:
-            service.render_project(project, outfile)
+            services.render_project(project, outfile)
 
         response_data = {
             "url": default_storage.url(path)
@@ -96,7 +96,7 @@ class ProjectImporterViewSet(mixins.ImportThrottlingPolicyMixin, CreateModelMixi
         total_memberships = len([m for m in data.get("memberships", [])
                                             if m.get("email", None) != data["owner"]])
         total_memberships = total_memberships + 1 # 1 is the owner
-        (enough_slots, error_message) = users_service.has_available_slot_for_import_new_project(
+        (enough_slots, error_message) = users_services.has_available_slot_for_import_new_project(
             self.request.user,
             is_private,
             total_memberships
@@ -105,22 +105,22 @@ class ProjectImporterViewSet(mixins.ImportThrottlingPolicyMixin, CreateModelMixi
             raise exc.NotEnoughSlotsForProject(is_private, total_memberships, error_message)
 
         # Create Project
-        project_serialized = service.store_project(data)
+        project_serialized = services.store.store_project(data)
 
         if not project_serialized:
-            raise exc.BadRequest(service.get_errors())
+            raise exc.BadRequest(services.store.get_errors())
 
         # Create roles
         roles_serialized = None
         if "roles" in data:
-            roles_serialized = service.store_roles(project_serialized.object, data)
+            roles_serialized = services.store.store_roles(project_serialized.object, data)
 
         if not roles_serialized:
             raise exc.BadRequest(_("We needed at least one role"))
 
         # Create memberships
         if "memberships" in data:
-            service.store_memberships(project_serialized.object, data)
+            services.store.store_memberships(project_serialized.object, data)
 
         try:
             owner_membership = project_serialized.object.memberships.get(user=project_serialized.object.owner)
@@ -137,57 +137,57 @@ class ProjectImporterViewSet(mixins.ImportThrottlingPolicyMixin, CreateModelMixi
 
         # Create project values choicess
         if "points" in data:
-            service.store_choices(project_serialized.object, data,
-                                  "points", serializers.PointsExportSerializer)
+            services.store.store_project_attributes_values(project_serialized.object, data,
+                                                           "points", serializers.PointsExportSerializer)
         if "issue_types" in data:
-            service.store_choices(project_serialized.object, data,
-                                  "issue_types",
-                                  serializers.IssueTypeExportSerializer)
+            services.store.store_project_attributes_values(project_serialized.object, data,
+                                                           "issue_types",
+                                                           serializers.IssueTypeExportSerializer)
         if "issue_statuses" in data:
-            service.store_choices(project_serialized.object, data,
-                                  "issue_statuses",
-                                  serializers.IssueStatusExportSerializer,)
+            services.store.store_project_attributes_values(project_serialized.object, data,
+                                                           "issue_statuses",
+                                                           serializers.IssueStatusExportSerializer,)
         if "us_statuses" in data:
-            service.store_choices(project_serialized.object, data,
-                                  "us_statuses",
-                                  serializers.UserStoryStatusExportSerializer,)
+            services.store.store_project_attributes_values(project_serialized.object, data,
+                                                           "us_statuses",
+                                                           serializers.UserStoryStatusExportSerializer,)
         if "task_statuses" in data:
-            service.store_choices(project_serialized.object, data,
-                                  "task_statuses",
-                                  serializers.TaskStatusExportSerializer)
+            services.store.store_project_attributes_values(project_serialized.object, data,
+                                                           "task_statuses",
+                                                           serializers.TaskStatusExportSerializer)
         if "priorities" in data:
-            service.store_choices(project_serialized.object, data,
-                                  "priorities",
-                                  serializers.PriorityExportSerializer)
+            services.store.store_project_attributes_values(project_serialized.object, data,
+                                                           "priorities",
+                                                           serializers.PriorityExportSerializer)
         if "severities" in data:
-            service.store_choices(project_serialized.object, data,
-                                  "severities",
-                                  serializers.SeverityExportSerializer)
+            services.store.store_project_attributes_values(project_serialized.object, data,
+                                                           "severities",
+                                                           serializers.SeverityExportSerializer)
 
         if ("points" in data or "issues_types" in data or
                 "issues_statuses" in data or "us_statuses" in data or
                 "task_statuses" in data or "priorities" in data or
                 "severities" in data):
-            service.store_default_choices(project_serialized.object, data)
+            services.store.store_default_project_attributes_values(project_serialized.object, data)
 
         # Created custom attributes
         if "userstorycustomattributes" in data:
-            service.store_custom_attributes(project_serialized.object, data,
-                                            "userstorycustomattributes",
-                                            serializers.UserStoryCustomAttributeExportSerializer)
+            services.store.store_custom_attributes(project_serialized.object, data,
+                                                   "userstorycustomattributes",
+                                                   serializers.UserStoryCustomAttributeExportSerializer)
 
         if "taskcustomattributes" in data:
-            service.store_custom_attributes(project_serialized.object, data,
-                                            "taskcustomattributes",
-                                            serializers.TaskCustomAttributeExportSerializer)
+            services.store.store_custom_attributes(project_serialized.object, data,
+                                                   "taskcustomattributes",
+                                                   serializers.TaskCustomAttributeExportSerializer)
 
         if "issuecustomattributes" in data:
-            service.store_custom_attributes(project_serialized.object, data,
-                                            "issuecustomattributes",
-                                            serializers.IssueCustomAttributeExportSerializer)
+            services.store.store_custom_attributes(project_serialized.object, data,
+                                                   "issuecustomattributes",
+                                                   serializers.IssueCustomAttributeExportSerializer)
 
         # Is there any error?
-        errors = service.get_errors()
+        errors = services.store.get_errors()
         if errors:
             raise exc.BadRequest(errors)
 
@@ -199,21 +199,33 @@ class ProjectImporterViewSet(mixins.ImportThrottlingPolicyMixin, CreateModelMixi
 
     @detail_route(methods=['post'])
     @method_decorator(atomic)
-    def issue(self, request, *args, **kwargs):
+    def milestone(self, request, *args, **kwargs):
         project = self.get_object_or_none()
         self.check_permissions(request, 'import_item', project)
 
-        signals.pre_save.disconnect(sender=Issue,
-                                    dispatch_uid="set_finished_date_when_edit_issue")
+        milestone = services.store.store_milestone(project, request.DATA.copy())
 
-        issue = service.store_issue(project, request.DATA.copy())
-
-        errors = service.get_errors()
+        errors = services.store.get_errors()
         if errors:
             raise exc.BadRequest(errors)
 
-        headers = self.get_success_headers(issue.data)
-        return response.Created(issue.data, headers=headers)
+        headers = self.get_success_headers(milestone.data)
+        return response.Created(milestone.data, headers=headers)
+
+    @detail_route(methods=['post'])
+    @method_decorator(atomic)
+    def us(self, request, *args, **kwargs):
+        project = self.get_object_or_none()
+        self.check_permissions(request, 'import_item', project)
+
+        us = services.store.store_user_story(project, request.DATA.copy())
+
+        errors = services.store.get_errors()
+        if errors:
+            raise exc.BadRequest(errors)
+
+        headers = self.get_success_headers(us.data)
+        return response.Created(us.data, headers=headers)
 
     @detail_route(methods=['post'])
     @method_decorator(atomic)
@@ -224,9 +236,9 @@ class ProjectImporterViewSet(mixins.ImportThrottlingPolicyMixin, CreateModelMixi
         signals.pre_save.disconnect(sender=Task,
                                     dispatch_uid="set_finished_date_when_edit_task")
 
-        task = service.store_task(project, request.DATA.copy())
+        task = services.store.store_task(project, request.DATA.copy())
 
-        errors = service.get_errors()
+        errors = services.store.get_errors()
         if errors:
             raise exc.BadRequest(errors)
 
@@ -235,33 +247,21 @@ class ProjectImporterViewSet(mixins.ImportThrottlingPolicyMixin, CreateModelMixi
 
     @detail_route(methods=['post'])
     @method_decorator(atomic)
-    def us(self, request, *args, **kwargs):
+    def issue(self, request, *args, **kwargs):
         project = self.get_object_or_none()
         self.check_permissions(request, 'import_item', project)
 
-        us = service.store_user_story(project, request.DATA.copy())
+        signals.pre_save.disconnect(sender=Issue,
+                                    dispatch_uid="set_finished_date_when_edit_issue")
 
-        errors = service.get_errors()
+        issue = services.store.store_issue(project, request.DATA.copy())
+
+        errors = services.store.get_errors()
         if errors:
             raise exc.BadRequest(errors)
 
-        headers = self.get_success_headers(us.data)
-        return response.Created(us.data, headers=headers)
-
-    @detail_route(methods=['post'])
-    @method_decorator(atomic)
-    def milestone(self, request, *args, **kwargs):
-        project = self.get_object_or_none()
-        self.check_permissions(request, 'import_item', project)
-
-        milestone = service.store_milestone(project, request.DATA.copy())
-
-        errors = service.get_errors()
-        if errors:
-            raise exc.BadRequest(errors)
-
-        headers = self.get_success_headers(milestone.data)
-        return response.Created(milestone.data, headers=headers)
+        headers = self.get_success_headers(issue.data)
+        return response.Created(issue.data, headers=headers)
 
     @detail_route(methods=['post'])
     @method_decorator(atomic)
@@ -269,9 +269,9 @@ class ProjectImporterViewSet(mixins.ImportThrottlingPolicyMixin, CreateModelMixi
         project = self.get_object_or_none()
         self.check_permissions(request, 'import_item', project)
 
-        wiki_page = service.store_wiki_page(project, request.DATA.copy())
+        wiki_page = services.store.store_wiki_page(project, request.DATA.copy())
 
-        errors = service.get_errors()
+        errors = services.store.get_errors()
         if errors:
             raise exc.BadRequest(errors)
 
@@ -284,9 +284,9 @@ class ProjectImporterViewSet(mixins.ImportThrottlingPolicyMixin, CreateModelMixi
         project = self.get_object_or_none()
         self.check_permissions(request, 'import_item', project)
 
-        wiki_link = service.store_wiki_link(project, request.DATA.copy())
+        wiki_link = services.store.store_wiki_link(project, request.DATA.copy())
 
-        errors = service.get_errors()
+        errors = services.store.get_errors()
         if errors:
             raise exc.BadRequest(errors)
 
@@ -327,7 +327,7 @@ class ProjectImporterViewSet(mixins.ImportThrottlingPolicyMixin, CreateModelMixi
         total_memberships = len([m for m in dump.get("memberships", [])
                                             if m.get("email", None) != dump["owner"]])
         total_memberships = total_memberships + 1 # 1 is the owner
-        (enough_slots, error_message) = users_service.has_available_slot_for_import_new_project(
+        (enough_slots, error_message) = users_services.has_available_slot_for_import_new_project(
             user,
             is_private,
             total_memberships
@@ -335,11 +335,23 @@ class ProjectImporterViewSet(mixins.ImportThrottlingPolicyMixin, CreateModelMixi
         if not enough_slots:
             raise exc.NotEnoughSlotsForProject(is_private, total_memberships, error_message)
 
+        # Async mode
         if settings.CELERY_ENABLED:
             task = tasks.load_project_dump.delay(user, dump)
             return response.Accepted({"import_id": task.id})
 
-        project = dump_service.dict_to_project(dump, request.user)
-        response_data = ProjectSerializer(project).data
-        return response.Created(response_data)
+        # Sync mode
+        try:
+            project = services.store_project_from_dict(dump, request.user)
+        except err.TaigaImportError as e:
+            # On Error
+            ## remove project
+            if e.project:
+                e.project.delete_related_content()
+                e.project.delete()
 
+            return response.BadRequest({"error": e.message, "details": e.errors})
+        else:
+            # On Success
+            response_data = ProjectSerializer(project).data
+            return response.Created(response_data)
