+++ taiga/export_import/services/__init__.py
@@ -15,21 +15,12 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from .development import *
+# This makes all code that import services works and
+# is not the baddest practice ;)
 
-CELERY_ENABLED = False
-CELERY_ALWAYS_EAGER = True
+from .render import render_project
+from . import render
 
-MEDIA_ROOT = "/tmp"
+from .store import store_project_from_dict
+from . import store
 
-EMAIL_BACKEND = "django.core.mail.backends.locmem.EmailBackend"
-INSTALLED_APPS = INSTALLED_APPS + [
-    "tests",
-]
-
-REST_FRAMEWORK["DEFAULT_THROTTLE_RATES"] = {
-    "anon": None,
-    "user": None,
-    "import-mode": None,
-    "import-dump-mode": None,
-}
