+++ taiga/export_import/tasks.py
@@ -27,10 +27,11 @@ from django.conf import settings
 from django.utils.translation import ugettext as _
 
 from taiga.base.mails import mail_builder
+from taiga.base.utils import json
 from taiga.celery import app
 
-from .service import render_project
-from .dump_service import dict_to_project
+from . import exceptions as err
+from . import services
 from .renderers import ExportRenderer
 
 logger = logging.getLogger('taiga.export_import')
@@ -46,7 +47,7 @@ def dump_project(self, user, project):
     try:
         url = default_storage.url(path)
         with default_storage.open(storage_path, mode="w") as outfile:
-            render_project(project, outfile)
+            services.render_project(project, outfile)
 
     except Exception:
         # Error
@@ -77,28 +78,57 @@ def delete_project_dump(project_id, project_slug, task_id):
     default_storage.delete("exports/{}/{}-{}.json".format(project_id, project_slug, task_id))
 
 
+ADMIN_ERROR_LOAD_PROJECT_DUMP_MESSAGE = _("""
+
+Error loading dump by {user_full_name} <{user_email}>:"
+
+
+REASON:
+-------
+{reason}
+
+DETAILS:
+--------
+{details}
+
+TRACE ERROR:
+------------""")
+
+
 @app.task
 def load_project_dump(user, dump):
     try:
-        project = dict_to_project(dump, user)
-    except Exception:
-        # Error
+        project = services.store_project_from_dict(dump, user)
+    except err.TaigaImportError as e:
+        # On Error
+        ## remove project
+        if e.project:
+            e.project.delete_related_content()
+            e.project.delete()
+
+        ## send email to the user
+        error_subject = _("Error loading project dump")
+        error_message = e.message or _("Error loading your project dump file")
+
         ctx = {
             "user": user,
-            "error_subject": _("Error loading project dump"),
-            "error_message": _("Error loading project dump"),
+            "error_subject": error_message,
+            "error_message": error_subject,
         }
         email = mail_builder.import_error(user, ctx)
         email.send()
-        logger.error('Error loading dump by %s <%s>',
-                     user,
-                     user.email,
-                     exc_info=sys.exc_info())
 
-        # TODO: [Rollback] Remove project because it can be corrupted
+        ## logged the error to sysadmins
+        text = ADMIN_ERROR_LOAD_PROJECT_DUMP_MESSAGE.format(
+            user_full_name=user,
+            user_email=user.email,
+            reason=e.message or _("  -- no detail info --"),
+            details=json.dumps(e.errors, indent=4)
+        )
+        logger.error(text, exc_info=sys.exc_info())
 
     else:
-        # Success
+        # On Success
         ctx = {"user": user, "project": project}
         email = mail_builder.load_dump(user, ctx)
         email.send()
