+++ tests/unit/test_export.py
@@ -20,7 +20,7 @@ import io
 from .. import factories as f
 
 from taiga.base.utils import json
-from taiga.export_import.service import render_project
+from taiga.export_import.services import render_project
 
 pytestmark = pytest.mark.django_db
